+++ lib/rubocop/cop/commissioner.rb
@@ -102,6 +102,11 @@ module RuboCop
         end
       end
 
+      # Allow blind rescues here, since we're absorbing and packaging or
+      # re-raising exceptions that can be raised from within the individual
+      # cops' `#investigate` methods.
+      #
+      # rubocop:disable Lint/RescueWithoutErrorClass
       def with_cop_error_handling(cop, node = nil)
         yield
       rescue => e
@@ -113,6 +118,7 @@ module RuboCop
         error = CopError.new(e, line, column)
         @errors[cop] << error
       end
+      # rubocop:enable Lint/RescueWithoutErrorClass
     end
   end
 end